C++ pipeline stage1: processors (#995)
* remove useless codes

* add UT

* Remove the limit that accelerate plugins cannot be used with other plugins

* fix accelerate plugin with only go flusher plugin

* fix UT

* fix member uninitialized

* fix pass to Go plugin tag name error

* fix c++ to go process

* fix process buffer legacy

* refine code according to comments

* stick to container naming convention, stick to smart pointer param passing convention

* fix split line if not enable_new_pipeline

+++ core/reader/LogFileReader.cpp
@@ -1926,7 +1926,7 @@ size_t LogFileReader::AlignLastCharacter(char* buffer, size_t size) {
             }
             endPs--;
         }
-        if (endPs - 1 + n >= size) {
+        if (endPs - 1 + n >= (int)size) {
             buffer[endPs] = '\0';
             return endPs;
         } else {
